API/Client refactoring
* client code split into separate files
* added skeleton for rspec tests
* updated examples
