+++ lib/occi/api/client_amqp.rb
@@ -1,0 +1,11 @@
+module Occi
+  module API
+    class ClientAmqp
+
+      def initialize(*args)
+        raise "Not yet implemented!"
+      end
+
+    end
+  end
+end
