+++ lib/occi/api/dsl.rb
@@ -2,8 +2,16 @@ module Occi
   module API
     module DSL
 
-      def connect(*args)
-        @client = Occi::API::Client.new(*args)
+      def connect(protocol, *args)
+
+        case protocol
+        when :http 
+          @client = Occi::API::ClientHttp.new(*args)
+        when :amqp
+          @client = Occi::API::ClientAmqp.new(*args)
+        else
+          raise "Protocol #{protocol.to_s} is not supported!"
+        end
 
         true
       end
