+++ spec/occi/client_spec.rb
@@ -2,11 +2,167 @@ require 'rspec'
 require 'occi'
 
 module Occi
-  describe Client do
+  module API
 
-    it "should do something" do
+    describe ClientHttp do
 
-      # TODO: Implement scenarios for client
+      it "should establish connection" do
+        # TODO
+      end
+
+      it "should list all available resources" do
+      end
+
+      it "should list compute resources" do
+      end
+
+      it "should list network resources" do
+      end
+
+      it "should list storage resources" do
+      end
+
+      it "should list all available mixins" do
+      end
+
+      it "should list os_tpl mixins" do
+      end
+
+      it "should list resource_tpl mixins" do
+      end
+
+      it "should describe all available resources" do
+      end
+
+      it "should describe compute resources" do
+      end
+
+      it "should describe network resources" do
+      end
+
+      it "should describe storage resources" do
+      end
+
+      it "should describe all available mixins" do
+      end
+
+      it "should describe os_tpl mixins" do
+      end
+
+      it "should describe resource_tpl mixins" do
+      end
+
+      it "should create a new compute resource" do
+      end
+
+      it "should create a new storage resource" do
+      end
+
+      it "should create a new network resource" do
+      end
+
+      it "should delete a compute resource" do
+      end
+
+      it "should delete a network resource" do
+      end
+
+      it "should delete a storage resource" do
+      end
+
+      it "should trigger an action on a compute resource" do
+      end
+
+      it "should trigger an action on a storage resource" do
+      end
+
+      it "should trigger an action on a network resource" do
+      end
+
+      it "should refresh its model" do
+      end
+
     end
+
+    describe ClientAmqp do
+
+      it "should establish connection" do
+        # TODO
+      end
+
+      it "should list all available resources" do
+      end
+
+      it "should list compute resources" do
+      end
+
+      it "should list network resources" do
+      end
+
+      it "should list storage resources" do
+      end
+
+      it "should list all available mixins" do
+      end
+
+      it "should list os_tpl mixins" do
+      end
+
+      it "should list resource_tpl mixins" do
+      end
+
+      it "should describe all available resources" do
+      end
+
+      it "should describe compute resources" do
+      end
+
+      it "should describe network resources" do
+      end
+
+      it "should describe storage resources" do
+      end
+
+      it "should describe all available mixins" do
+      end
+
+      it "should describe os_tpl mixins" do
+      end
+
+      it "should describe resource_tpl mixins" do
+      end
+
+      it "should create a new compute resource" do
+      end
+
+      it "should create a new storage resource" do
+      end
+
+      it "should create a new network resource" do
+      end
+
+      it "should delete a compute resource" do
+      end
+
+      it "should delete a network resource" do
+      end
+
+      it "should delete a storage resource" do
+      end
+
+      it "should trigger an action on a compute resource" do
+      end
+
+      it "should trigger an action on a storage resource" do
+      end
+
+      it "should trigger an action on a network resource" do
+      end
+
+      it "should refresh its model" do
+      end
+
+    end
+
   end
 end
